Move components to page directories
