+++ client/src/archipelago-layout/list/List.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { List as RaList } from 'react-admin';
-import {ResourceTabs} from "../index";
+import ResourceTabs from "./ResourceTabs";
 
 const List = ({ children, ...otherProps }) => (
   <RaList perPage={50} {...otherProps}>
