+++ client/src/archipelago-layout/show/Column/RightLabel.js
