+++ client/src/archipelago-layout/show/ColumnShowLayout.js
