+++ client/src/archipelago-layout/show/Hero/DetailsList.js
